atmospheric free flight bugfix
A hold-prograde flight mode is quite tricky: flight angles are different when measured in a stationary frame of reference (orbital) and rotating one (surface). Pitch is calculated okay in both frames, but yaw is trickier. For now I will lock yaw to precalculated launch azimuth instead of air-relative angle, which might(?) introduce some AoA in an actual flight. This seems to be the most straightforward way to kill a bug that occurs now: vehicle aligns to a wrong angle and PEG can't reach target.