events: sort out the chaos between UPFG control, staging, and UI
This introduces one big change to how things are handled: previously,
the flag `stagingInProgress` was FALSE during the time between exiting
the passive guidance loop and engaging UPFG for the first time. From
now on, this flag will be TRUE whenever any transition between actively
guided happens - including the first one.

Additionally, several extra flags to explicitly convey information
about the current state of the vehicle and guidance are added:
* `activeGuidanceMode`
* `prestageHold`
* `upfgEngaged`
* `"isSustainer"`, internal to each vehicle stage.
None of those introduce any new concept, it's just giving clear names
to existing concepts and rewriting some code to use those concepts in
a transparent fashion, as well as updating the UI to present those
concepts to the user.
Consult upfgSteeringControl for a more in-depth explanation.