actually kill steering at terminal guidance
Previously, terminal guidance was supposed to maintain the last good
output of UPFG. But if staging occurred just before transition and UPFG
recalculated guidance, producing a different attitude, PEGAS would
attempt to reorient the vehicle. For a sufficiently heavy vehicle and
sufficient attitude different, this could lead to a sharp rotation,
possibly too large to be completed in `upfgFinalizationTime` seconds.
This would leave the vehicle tumbling after the terminal loop exited.
Now steering is set to explicitly kill any rotation, even if the last
attitude was not very "good". If you're having any issues with this
part of the flight, you're probably trying to stage too close to MECO.